requirements.txt
streamlit
python-dotenv
langchain
langchain_groq
langchain_objectbox
langchain_community
PyPDF2
ollama
objectbox
